Converted from darcs to Git.
The commits on the master branch use the same (linear) ordering as the
corresponding patches used in my master darcs repository.

Darcs does not seem to record information about the local time zone in
effect when a patch is recorded. I did not try to recreate this
information, all commits before this one have UTC as their time zone.